Replace current routing with React-Router's Outlet component. Create a 'Layout.jsx' component. Rename the 'RecurringTrans.jsx' component to 'RecurringTransactions.jsx'.
